server/requirements.txt
aiohttp==3.*
python-dotenv==1.*

# pip install -r requirements.txt
